requirements.txt
boto3                      
sagemaker        
torch                 
librosa                   
timm                    
einops                    
nnAudio               
python-dotenv             

